refactor: extract createApp factory to eliminate code duplication
- Created packages/backend/src/app.ts with createApp() factory function
- Consolidates Express app setup (body parsing, routing-controllers, middleware)
- src/index.ts now uses createApp() instead of inline config
- tests/test-helpers.ts now uses createApp() instead of duplicating config
- Single source of truth for app configuration
- Both production and test environments use identical app setup
- Improves maintainability: changes to app config only need to be made once
